+++ components/sections/WhoWeAre.jsx
@@ -1,4 +1,4 @@
-import { DefaultButton } from "../ui/Button";
+import DefaultButton from "../ui/DefaultButton";
 import { Heading5 } from "../ui/Headings";
 import Title from '../ui/Title'
 const WhoWeAre = () => {
