+++ components/ui/LargeButton.jsx
@@ -1,3 +1,6 @@
+import { Arrow } from "../svgs/Arrow";
+import Link from "next/link";
+
 const LargeButton = ({ text, hasArrow, styleSpan, style, hasLink }) => {
     return ( 
         <button
